Extract all time series from a gts object

The time series from all levels of a hierarchical/grouped time series or a forecasted hierarchical/grouped time series are returned as a multivariate time series.

allts(y, forecasts = TRUE)

Arguments

y
An object of class gts.

forecasts
If y contains forecasts and historical data, then forecasts indicates whether to return the forecasts or the historical data. Otherwise it is ignored.

See also

aggts

Examples

allts(htseg1)

#> Time Series:
#> Start = 1992 
#> End = 2001 
#> Frequency = 1 
#>         Total        A        B       AA        AB       AC       BA       BB
#> 1992 48.74808 27.21134 21.53674 7.862669  9.267647 10.08102 10.49276 11.04398
#> 1993 49.48047 27.83883 21.64164 8.340355  9.416544 10.08193 10.51712 11.12452
#> 1994 49.93238 28.14535 21.78704 8.549132  9.479022 10.11719 10.54013 11.24691
#> 1995 50.24070 28.27712 21.96358 8.597699  9.532376 10.14705 10.61220 11.35138
#> 1996 50.60846 28.47800 22.13046 8.770196  9.547591 10.16021 10.64932 11.48114
#> 1997 50.84851 28.56447 22.28404 8.837793  9.555664 10.17101 10.70736 11.57668
#> 1998 51.70922 28.90753 22.80169 8.893267  9.730347 10.28392 10.86323 11.93846
#> 1999 51.94330 29.02155 22.92175 8.933993  9.797294 10.29026 10.92463 11.99712
#> 2000 52.57796 29.40308 23.17488 9.133653  9.921262 10.34817 11.00622 12.16865
#> 2001 53.21496 29.64248 23.57248 9.152636 10.022783 10.46706 11.03531 12.53717